% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-af_crime93.R
\docType{data}
\name{af_crime93}
\alias{af_crime93}
\title{Statewide Crime Data (1993)}
\format{
A data frame with 51 observations on the following 8 variables.
\describe{
\item{\code{state}}{a character vector for the state}
\item{\code{violent}}{a numeric vector for the violent crime rate (per 100,000 people in population)}
\item{\code{murder}}{a numeric vector for the murder rate (per 100,000 people in population)}
\item{\code{poverty}}{a numeric vector for the  percent with income below the poverty level}
\item{\code{single}}{a numeric vector for the percent of families headed by a single parent}
\item{\code{metro}}{a numeric vector for the percent of population in metropolitan areas}
\item{\code{white}}{a numeric vector for the percentage of the state that is white}
\item{\code{highschool}}{a numeric vector for the percent of state that graduated from high school}
}
}
\usage{
af_crime93
}
\description{
These data are in Table 9.1 of the 3rd edition of Agresti and Finlay's
 *Statistical Methods for the Social Sciences*. The data are from
 *Statistical Abstract of the United States* and most variables were measured in 1993.
}
\details{
The data are from Statistical Abstract of the United States and most variables were measured in 1993. These data should
result in regressions that would flunk a Breusch-Pagan test for heteroskedasticity.
}
\references{
Agresti, Alan and Barbara Finley. 1997. \emph{Statistical Methods for the Social Sciences}. Prentice Hall. (3rd Edition)
}
\keyword{datasets}
