% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-aluminum_premiums.R
\docType{data}
\name{aluminum_premiums}
\alias{aluminum_premiums}
\title{LME Aluminum Premiums Data}
\format{
A data frame with 2,812 observations on the following 3 variables.
\describe{
\item{\code{date}}{a date}
\item{\code{group}}{a factor with levels of \code{East Asia}, \code{Southeast Asia},
\code{United States}, and \code{Western Europe}}
\item{\code{price}}{a numeric vector for the price of the LME aluminum premium}
}
}
\usage{
aluminum_premiums
}
\description{
A near daily data set on the price of aluminum premiums (USD/MT) for LME
 in the U.S., Western Europe, East Asia, and Southeast Asia. I like these
 data as illustrative of some of the shortsightedness of the aluminum tariffs
 that Donald Trump announced in March 2018. The tariffs had no discernible
 effect on manufacturing employment or earnings, but they created a supply shock
 that made aluminum more expensive.
}
\details{
LME aluminum premiums (monthly contracts going out to 15 months) work
alongside LME aluminum contracts to allow market participants to hedge the all-in
price and physically deliver or receive premium aluminum warrants in
non-queued LME premium warehouses.
}
\keyword{datasets}
