% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-Arca.R
\docType{data}
\name{Arca}
\alias{Arca}
\title{NYSE Arca Steel Index data, 2017–2020}
\format{
A data frame with 966 observations on the following 6 variables.
\describe{
\item{\code{date}}{the date}
\item{\code{close}}{the closing price}
\item{\code{open}}{the opening price}
\item{\code{high}}{the daily high in that day's trading}
\item{\code{low}}{the daily low in that day's trading}
}
}
\usage{
Arca
}
\description{
Daily data on the NYSE Arca Steel Index. These data are useful for me in
teaching how Trump's 2018 steel tariffs didn't do much good for the steel industry.
}
\details{
These data are taken from \code{investing.com}.
Code name: \code{STTL}.
}
\keyword{datasets}
