% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-arcticseaice.R
\docType{data}
\name{arcticseaice}
\alias{arcticseaice}
\title{Arctic Sea Ice Extent Data, 1901-2015}
\format{
A data frame with 115 observations on the following 4 variables.
\describe{
\item{\code{year}}{the year}
\item{\code{value}}{the annual Arctic sea ice extent (in 10^6 sq km)}
\item{\code{ub}}{The upper bound of the value, provided by Connelly et al.}
\item{\code{lb}}{The lower bound of the value, provided by Connelly et al.}
}
}
\usage{
arcticseaice
}
\description{
This data set from Connelly et al. (2017) measures the Arctic sea
ice extent in 10^6 square kilometers. It includes lower bounds and
upper bounds on annual averages.
}
\details{
This is for illustration of climate change for my intro students.
Connelly et al. (2017) are in part a methodological paper. The data I present
here are from the "rescaled (unadjusted T)" data in the second sheet from
their replication files.
}
\references{
Connolly et al. (2017), ''Re-calibration of Arctic sea ice extent
datasets using Arctic surface air temperature records''.
\emph{Hydrological Sciences Journal} 62(8): 1317--40.
}
\keyword{datasets}
