% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-arg_tariff.R
\docType{data}
\name{arg_tariff}
\alias{arg_tariff}
\title{Simple Mean Tariff Rate for Argentina}
\format{
A data frame with three variables:
\describe{
\item{\code{country}}{country name (Argentina)}
\item{\code{year}}{the year}
\item{\code{tariffrate}}{the simple mean tariff rate for Argentina on all products (as a percentage)}
}
}
\usage{
arg_tariff
}
\description{
Simple mean tariff rate for Argentina, starting in 1980.
The goal is to keep these data current.
}
\details{
Data come from various sources. World Bank estimates are used for
1980-1984 and 2010-2018, but see also Lora's (2012) report for the Inter-American
Development Bank. The 1980-1984 estimates are actually means for 1980-1 and 1982-4 via
Laird and Nogues' (1989) article in the World Bank Economic Review.
}
\keyword{datasets}
