% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-co2emissions.R
\docType{data}
\name{co2emissions}
\alias{co2emissions}
\title{Carbon Dioxide Emissions Data}
\format{
A data frame with 3,099 observations on the following 2 variables.
\describe{
\item{\code{year}}{the year (negative values = BCE)}
\item{\code{value}}{estimated carbon dioxide emissions (in ppm)}
}
}
\usage{
co2emissions
}
\description{
This is a sample data set, cobbled from various sources, about
 carbon dioxide emissions in the history of the planet from 800,000 BCE to
 the most recently concluded calendar year. I use this for a data visualization
 example for a lecture on climate change and international politics.
 Data communicate yearly averages/estimates.
}
\details{
The data come from many sources. Before 0 CE, the data come from 10 sources
described here by the EPA
(\url{https://www.epa.gov/climate-indicators/climate-change-indicators-atmospheric-concentrations-greenhouse-gases}).
Observations from 0 CE to 2014 come from Meinshausen et al. (2017)
(\url{https://gmd.copernicus.org/articles/10/2057/2017/}). Observations from 2015 forward come
from NASA (\url{https://climate.nasa.gov/vital-signs/carbon-dioxide/}).
}
\references{

}
\keyword{datasets}
