% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-coffee_imports.R
\docType{data}
\name{coffee_imports}
\alias{coffee_imports}
\title{Coffee Imports for Select Importing Countries}
\format{
A data frame with 1050 observations on the following 3 variables.
\describe{
\item{\code{country}}{a character vector for the country}
\item{\code{year}}{a numeric vector for the year}
\item{\code{value}}{a numeric vector for the coffee imports for all select importing countries (in thousand 60-kg bags)}
}
}
\usage{
coffee_imports
}
\description{
A simple panel on coffee imports for select importing countries
(i.e. European Union + Japan + Norway + Russia + Switzerland + Tunisia +
the United Kingdom + United States).
}
\details{
Data come from the International Coffee Organization, of which I feel I should be a member.
}
\keyword{datasets}
