% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-coffee_price.R
\docType{data}
\name{coffee_price}
\alias{coffee_price}
\title{The Primary Commodity Price for Coffee (Arabica, Robustas)}
\format{
A data frame with the following 3 variables.
\describe{
\item{\code{date}}{the date (year-month)}
\item{\code{arabica}}{the price (monthly average) of mild Arabica, via
International Coffee Organization data, in nominal US cents per pound}
\item{\code{robustas}}{the price (monthly average) of Robustas,
via International Coffee Organization data, in nominal US cents per pound}
}
}
\usage{
coffee_price
}
\description{
This is primary commodity price data for coffee (Arabica, Robustas)
from 1980 to the present. I manually update these data since FRED's
coverage since 2017 has been spotty.
}
\details{
Data come from International Monetary Fund (Primary Commodity Prices) and
International Coffee Organization. The IMF adds these prices are global and the
New York cash price, ex-dock
}
\keyword{datasets}
