% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-country_isocodes.R
\docType{data}
\name{country_isocodes}
\alias{country_isocodes}
\title{ISO 3166 Country Codes (Two-Character, Three-Character, Numeric)}
\format{
A data frame with 249 observations on the following 4 variables.
\describe{
\item{\code{iso2c}}{a two-character ISO code}
\item{\code{iso3c}}{a three-character ISO code}
\item{\code{iso3n}}{a three-digit numeric ISO code}
\item{\code{name}}{an English country name}
}
}
\usage{
country_isocodes
}
\description{
A data set of country ISO codes, for my ease and for the ease of my students.
}
\details{
This is a simple, abbreviated port and rename of the
\code{ISO3_166_1} data in the \pkg{ISOcodes} package.
}
\keyword{datasets}
