% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-eq_passengercars.R
\docType{data}
\name{eq_passengercars}
\alias{eq_passengercars}
\title{Export Quality Data for Passenger Cars, 1963-2014}
\format{
A data frame with 60424 observations on the following 6 variables.
\describe{
\item{\code{country}}{a character vector for the country/area.}
\item{\code{ccode}}{a numeric vector for the Correlates of War country code.}
\item{\code{category}}{a factor with levels \code{Export Quality Index},
 \code{Export quality 95 percent interval - lower bound},
 \code{Export quality 95 percent interval - upper bound}
 \code{Unit value of exports}, \code{Unit value 95 percent interval - lower bound},
 \code{Unit value 95 percent interval - upper bound},
 \code{Trade value of exports}}
\item{\code{type}}{a factor with levels \code{51. Transport equipment, Passenger cars}.
This is a constant. I just felt like making it a factor.}
\item{\code{year}}{a numeric vector for the year}
\item{\code{value}}{a numeric vector for the value of the particular category.}
}
}
\usage{
eq_passengercars
}
\description{
Data from the International Monetary Fund for the export quality and
 unit/trade value of passenger cars for all available countries and years
 from 1963 to 2014.
}
\keyword{datasets}
