% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-eustates.R
\docType{data}
\name{eustates}
\alias{eustates}
\title{EU Member States (Current as of 2019)}
\format{
A data frame with 28 observations on the following 3 variables.
\describe{
\item{\code{date}}{a date indicating accession}
\item{\code{country}}{a character vector for the country}
\item{\code{iso2c}}{a character vector for iso2c}
}
}
\usage{
eustates
}
\description{
European Union membership by accession date
}
\details{
Data come from the European Union's website.
}
\keyword{datasets}
