% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-gss_abortion.R
\docType{data}
\name{gss_abortion}
\alias{gss_abortion}
\title{Abortion Opinions in the General Social Survey (1972-2018)}
\format{

}
\usage{
gss_abortion
}
\description{
This is a toy data set derived from the General Social Survey that I intend
to use for several purposes. First, the battery of abortion items can serve as
toy data to illustrate mixed effects modeling as equivalent to a
one-parameter (Rasch) model. Second, I include some covariates to also do some
basic regressions. I think abortion opinions are useful learning tools for
statistical inference for college students.
}
\details{
Data include all General Social Survey observations from 1972 to 2018 for
these variables. Be mindful of missing data.
}
\keyword{datasets}
