% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-gss_wages.R
\docType{data}
\name{gss_wages}
\alias{gss_wages}
\title{The Gender Pay Gap in the General Social Survey}
\format{

}
\usage{
gss_wages
}
\description{
Wage data from the General Social Survey (1974-2018) to illustrate
 wage discrepancies by gender (while also considering respondent occupation, age, and education).
}
\details{
For further details, see \url{https://gssdataexplorer.norc.org}. Consult \url{https://census.gov} for more information about occupation codes.
}
\keyword{datasets}
