% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-min_wage.R
\docType{data}
\name{min_wage}
\alias{min_wage}
\title{History of Federal Minimum Wage Rates Under the Fair Labor Standards Act, 1938-2009}
\format{
A data frame with 23 observations on the following 5 variables.
\describe{
\item{\code{date}}{a date for when a new minimum wage was introduced}
\item{\code{wage}}{the (nominal) value of the wage}
}
}
\source{
\url{https://www.dol.gov/agencies/whd/minimum-wage/history/chart}
}
\usage{
min_wage
}
\description{
A data set on the various federal minimum wage rates.
}
\details{
Data come from the Department of Labor. Wages are taken from wage adjustments
from the 1938 act.
}
\keyword{datasets}
