% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-Newhouse77.R
\docType{data}
\name{Newhouse77}
\alias{Newhouse77}
\title{Medical-Care Expenditure: A Cross-National Survey (Newhouse, 1977)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 5 columns.
}
\usage{
Newhouse77
}
\description{
These are the data in Newhouse's (1977) simple OLS model from 1977. In
 his case, he's trying to explain medical care expenditures as a function of GDP
 per capita for these countries. It's probably the easiest OLS model I can find
 in print because Newhouse helpfully provides all the data in one simple table.
}
\details{
Table 1 in Newhouse (1977) is well-annotated with background information.
}
\references{
Newhouse, Joseph P. 1977. "Medical-Care Expenditure: A Cross-National Survey." \emph{Journal of Human Resources} 12(1): 115-125.
}
\keyword{datasets}
