% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-Presidents.R
\docType{data}
\name{Presidents}
\alias{Presidents}
\title{U.S. Presidents and Their Terms in Office}
\format{
A data frame with 45 observations on the following 3 variables.
\describe{
\item{\code{president}}{the president}
\item{\code{start}}{the start date of the term, as a date}
\item{\code{end}}{the end date of the term, as a date}
}
}
\usage{
Presidents
}
\description{
This should be self-evident. Here are all U.S. presidents who have completed
their terms in office (i.e. excluding the current one).
}
\details{
I scraped this from \url{https://www.presidentsusa.net}.
Data frame is capital-P "Presidents" to avoid a conflict with the
\code{presidents} data frame from the \code{datasets} package.
}
\keyword{datasets}
