% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-pwt_sample.R
\docType{data}
\name{pwt_sample}
\alias{pwt_sample}
\title{Penn World Table (10.0) Macroeconomic Data for Select Countries, 1950-2019}
\format{
A data frame with 1470 observations on the following 11 variables.
\describe{
\item{\code{country}}{the country name}
\item{\code{isocode}}{The country's ISO code}
\item{\code{year}}{a numeric vector for the year}
\item{\code{pop}}{Population in millions}
\item{\code{hc}}{Index of human capital per person, based on years of schooling and returns to education}
\item{\code{rgdpna}}{Real GDP at constant 2011 national prices (in million 2017 USD)}
\item{\code{rgdpo}}{Output-side real GDP at chained PPPs (in million 2017 USD)}
\item{\code{rgdpe}}{Expenditure-side real GDP at chained PPPs (in million 2017 USD)}
\item{\code{labsh}}{Share of labor compensation in GDP at current national prices}
\item{\code{avh}}{Average annual hours worked by persons engaged.}
\item{\code{emp}}{Number of persons engaged (in millions)}
\item{\code{rnna}}{Capital stock at constant 2017 national prices (in million 2017 USD)}
}
}
\source{
Taken from the \code{pwt10} package. See: \url{https://www.rug.nl/ggdc/}
}
\usage{
pwt_sample
}
\description{
These are some macroeconomic data for 21 select (rich) countries. I've used
these data before to discuss issues of grouping and skew in cross-sectional data.
}
\keyword{datasets}
