% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-quartets.R
\docType{data}
\name{quartets}
\alias{quartets}
\title{Anscombe's (1973) Quartets}
\format{
A data frame with 44 observations on the following 3 variables.
\describe{
\item{\code{group}}{a categorical identifier for the quartet}
\item{\code{x}}{a continuous variable}
\item{\code{y}}{a continuous variable}
}
}
\usage{
quartets
}
\description{
These are four x-y data sets, combined into a long format, which have the
 same traditional statistical properties (mean, variance, correlation,
 regression line, etc.). However, they look quite different.
}
\details{
Data come default in R, but I elected to change the format to be
a bit more accessible.
}
\references{
Anscombe, Francis J. (1973). "Graphs in Statistical Analysis." \emph{The American Statistician} 27: 17–21.
}
\keyword{datasets}
