% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-SBCD.R
\docType{data}
\name{SBCD}
\alias{SBCD}
\title{Systemic Banking Crises Database II}
\format{
A data frame with 574 observations on the following 4 variables.
\describe{
\item{\code{country}}{the country, as it appears in the data}
\item{\code{type}}{the type of crisis, entered here as "banking", "currency", "debt", or "debtrestructuring"}
\item{\code{year}}{the year of the crisis}
\item{\code{month}}{the month the crisis started, if known}
}
}
\usage{
SBCD
}
\description{
A data set on banking, currency, debt, and debt-restructuring crises from
1970 to 2017.
}
\details{
Data are cobbled from the second and third sheets of the spreadsheet
the authors provide. Country names are as entered in their spreadsheet.
Liberia has an "NA" in the raw data for sovereign debt restructuring and I
don't know why. I elect to keep it.
}
\references{
Laeven, Luc and Fabian Valencia. 2020. "Systemic Banking Crises
Database II". \emph{IMF Economic Review} 68: 307--361.
}
\keyword{datasets}
