% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-SCP16.R
\docType{data}
\name{SCP16}
\alias{SCP16}
\title{South Carolina County GOP/Democratic Primary Data, 2016}
\format{
A data frame with 46 observations on the following 15 variables.
\describe{
\item{\code{county}}{the county}
\item{\code{clinton}}{Hillary Clinton's county-level vote share in the 2016 party primary}
\item{\code{sanders}}{Bernie Sanders' county-level vote share in the 2016 party primary}
\item{\code{trump}}{Donald Trump's county-level vote share in the 2016 party primary}
\item{\code{cruz}}{Ted Cruz' county-level vote share in the 2016 party primary}
\item{\code{rubio}}{Marco Rubio's county-level vote share in the 2016 party primary}
\item{\code{percapinc}}{A county-level estimate for per capita income}
\item{\code{medhouseinc}}{A county-level estimate for the median household income}
\item{\code{medfaminc}}{A county-level estimate for the median family income}
\item{\code{illiteracy}}{An estimate of the percent of the county lacking "basic" prose literacy skills}
\item{\code{perblack}}{Percentage of the county that is black}
\item{\code{population}}{An estimate of the county-level population}
\item{\code{romneyshare2012}}{Mitt Romney's vote share at the county-level from the 2012 general election}
\item{\code{perhsgrad}}{Percentage of the county whose residents 25 years and older have at least a high school education}
\item{\code{unemployment}}{Unemployment rate for the county for January 2016}
}
}
\usage{
SCP16
}
\description{
County-level data on vote share and various background/demographic
 information for the 2016 South Carolina GOP/Democratic primaries.
}
\details{
The illiteracy estimate comes from a Department of Education report from 2003.
The unemployment rate data come from the Bureau of Labor Statistics. A Github repository contains
more information: \url{https://github.com/svmiller/sc-primary-2016}.
}
\keyword{datasets}
