% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-shuparks.R
\docType{data}
\name{shuparks}
\alias{shuparks}
\title{The Effect of Urban Parks on the Thermal Environment in Shanghai}
\format{
A data frame with 24 observations on the following 10 variables.
\describe{
\item{\code{group}}{a grouping category for park size, either 'Super Large', 'Large', 'Medium', or 'Small'}
\item{\code{name}}{the name of the park}
\item{\code{ha}}{park area, in hectares}
\item{\code{areap}}{percentage of the area of each park in the built-up part of Shanghai}
\item{\code{gc}}{the percentage of the trees or grass area in each park divided by the park area}
\item{\code{func}}{the supposed function of each park, per the Shanghai Landscape and City Appearance Administrative Bureau and Forestry Bureau}
\item{\code{mcd15}}{the maximum cooling distance, in meters, in 2015}
\item{\code{mci15}}{the maximum cooling intensity, in Celsius, in 2015}
\item{\code{mcd20}}{the maximum cooling distance, in meters, in 2020}
\item{\code{mci20}}{the maximum cooling intensity, in Celsius, in 2020}
}
}
\usage{
shuparks
}
\description{
A data set on urban parks in Shanghai and their effect on temperature.
}
\details{
Data come from Table 1 and Table 5 in Wang et al. (2022). Read that article
for more technical details.
}
\references{
Wang, Tian, Gui Tu, Bo Min, Zuzheng Li, Xiaofang Li, and Qingxiang You. 2022.
"The Mitigation Effect of Park Landscape on Thermal Environment in
Shanghai City Based on Remote Sensing Retrieval Method." \emph{International Journal
of Environmental Research and Public Health} 19(5): 2949.
}
\keyword{datasets}
