% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-steves_clothes.R
\docType{data}
\name{steves_clothes}
\alias{steves_clothes}
\title{Steve's (Professional) Clothes, as of March 3, 2019}
\format{
A data frame with 79 observations on the following 4 variables.
\describe{
\item{\code{type}}{Type of clothing}
\item{\code{brand}}{The brand of clothing (e.g. Apt. 9, Saddlebred)}
\item{\code{color}}{the color (and/or pattern) of the article of clothing}
\item{\code{origin}}{The country that produced the garment.}
}
}
\source{
Steve's closet. Hey, that's me!
}
\usage{
steves_clothes
}
\description{
I cobbled together this data set of the professional clothes
 (polos, long-sleeve dress shirts, pants) in my closet, largely for
 illustration on the origins of apparel in the U.S. for an intro lecture on trade.
}
\details{
If you must know, I do most of my clothes shopping at major retailers
in the U.S. (mostly Belk, J.C. Penney, and Kohl's). If that's you as well, the
odds are good the distribution of my clothes will closely resemble yours.)
}
\keyword{datasets}
