% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-stevesteps.R
\docType{data}
\name{stevesteps}
\alias{stevesteps}
\title{Steve's Steps in a Day, 2017-2025}
\format{
A data frame with 2875 observations on the following 2 variables.
\describe{
\item{\code{date}}{a date}
\item{\code{value}}{estimated steps in a day}
}
}
\usage{
stevesteps
}
\description{
These are data from my Fitbit for steps in a day.
}
\keyword{datasets}
