% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-sugar_price.R
\docType{data}
\name{sugar_price}
\alias{sugar_price}
\title{IMF Primary Commodity Price Data for Sugar}
\format{
A data frame with 1,316 observations on the following 3 variables.
\describe{
\item{\code{date}}{a date}
\item{\code{category}}{the category (either the U.S., global, or Europe)}
\item{\code{value}}{the price of sugar in U.S. cents per pound (NSA, nominal)}
}
}
\source{
International Monetary Fund
}
\usage{
sugar_price
}
\description{
This is primary commodity price data for sugar globally,
 in the United States, and in Europe for every month from 1980 to (roughly)
 the present. Prices are nominal U.S. cents per pound and are
 not seasonally adjusted ("NSA").
}
\details{
The price data for Europe do not appear to be updated as regularly as the
global and U.S. prices. Thus, the last month in the data for Europe are June 2017.
For that reason, I elected to make a data set of these data for posterity's sake.
}
\keyword{datasets}
