% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-usa_computers.R
\docType{data}
\name{usa_computers}
\alias{usa_computers}
\title{Percentage of U.S. Households with Computer Access, by Year}
\format{
A data frame with 19 observations on the following 2 variables.
\describe{
 \item{\code{year}}{the year}
 \item{\code{value}}{the estimated percentage of households with access to a computer}
}
}
\source{
Various: U.S. Census Bureau, Current Population Survey, and
American Community Survey
}
\usage{
usa_computers
}
\description{
This is a simple and regrettably incomplete time-series on the percentage
 of U.S. households with access to a computer, by year.
}
\details{
Data are spotty and regrettably this is not a perfect time-series.
However, it is useful for an in-class exercise to show that the proliferation of
household computers (over time) in the United States comes in part because of
globalization. Use it for that purpose. The data are reasonably faithful, but don't
treat it as gospel. Exact sourcing available upon request.
}
\keyword{datasets}
