% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-usa_states.R
\docType{data}
\name{usa_states}
\alias{usa_states}
\title{State Abbreviations, Names, and Regions/Divisions}
\format{
A data frame with 51 observations on the following 4 variables.
\describe{
\item{\code{stateabb}}{the state abbreviation}
\item{\code{statename}}{the state's name}
\item{\code{region}}{the state's Census region}
\item{\code{division}}{the state's Census division}
}
}
\usage{
usa_states
}
\description{
A simple data set from \code{state.abb}, \code{state.name}, \code{state.region}, and
\code{state.division} (+ District of Columbia). I'd rather just have all these
in one place.
}
\keyword{datasets}
