% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-wbd_example.R
\docType{data}
\name{wbd_example}
\alias{wbd_example}
\title{A Simple Panel drawn from World Bank Open Data}
\format{
A data frame with 4537 observations on the following 7 variables.
\describe{
\item{\code{country}}{an English name for the country/territorial unit}
\item{\code{iso2c}}{the two-character ISO code for the country/territorial unit}
\item{\code{iso3c}}{the three-character ISO code for the country/territorial unit}
\item{\code{year}}{the year of observation}
\item{\code{rgdppc}}{the real GDP per capita of the country/territorial unit in that year}
\item{\code{lifeexp}}{the average life expectancy at birth for men and women for the country that year}
\item{\code{hci}}{the human capital index [0-1] for the country that year}
}
}
\usage{
wbd_example
}
\description{
A simple data set drawn from World Bank Open Data. I'll use it to illustrate
some merge issues you might encounter in panel data.
}
\details{
The idea for this data comes by way of a student encounter where we noticed
this issue. Data were further generated by the wonderful \pkg{WDI} package.
The underlying data come from the World Bank national accounts (GDP per
capita), World Bank analyst estimates (human capital index), or the United
Nations Population Division (life expectancy at birth).
}
\keyword{datasets}
