% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-yugo_sales.R
\docType{data}
\name{yugo_sales}
\alias{yugo_sales}
\title{Yugo Sales in the United States, 1985-1992}
\format{
A data frame with 24 observations on the following 3 variables.
\describe{
\item{\code{year}}{the year}
\item{\code{car}}{the car type, either the Hyundai Excel, Yugo, or Toyota Tercel}
\item{\code{sales}}{the number of units sold in the United States}
}
}
\usage{
yugo_sales
}
\description{
A data set on Yugo sales against two competing models in the United States
 from 1985 to 1992.
}
\details{
Data come from a website then known as \code{carsalesbase.com}. I'm
aware the inclusion of the Tercel is questionable since the third generation
of Tercels were quite different from the first and second generations.
However, I use these data to illustrate how poorly the Yugo fared against
competing models, including the first and second generation Tercels. I think
the inclusion is fair for that purpose.
}
\keyword{datasets}
