% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-coffee_imports.R
\docType{data}
\name{coffee_imports}
\alias{coffee_imports}
\title{Coffee Imports for Select Importing Countries}
\format{
A data frame with 29 observations on the following 3 variables.
\describe{
\item{\code{year}}{the year}
\item{\code{imports}}{coffee imports for all select importing countries (in thousand 60-kg bags)}
\item{\code{usaimports}}{coffee imports for just the United States (in thousand 60-kg bags)}
}
}
\usage{
coffee_imports
}
\description{
A simple time series on coffee imports for select importing
 countries (i.e. European Union + Japan + Russia + Tunisia + United States).
}
\details{
Data come from the International Coffee Organization, of which I feel I should be a member.
}
\keyword{datasets}
